+++ saml-query-profile-client/src/main/java/org/wso2/carbon/identity/saml/query/profile/test/SAMLAssertionIDRequestClient.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.query.saml.test;
+package org.wso2.carbon.identity.saml.query.profile.test;
 
 import org.apache.axiom.om.OMElement;
 import org.apache.axis2.client.ServiceClient;
