+++ saml-query-profile-client/src/main/java/org/wso2/carbon/identity/saml/query/profile/test/SAMLAttributeQueryRequestClient.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.query.saml.test.invalid;
+package org.wso2.carbon.identity.saml.query.profile.test;
 
 import org.apache.axiom.om.OMElement;
 import org.apache.axis2.client.ServiceClient;
@@ -39,17 +39,15 @@ import org.opensaml.saml.saml2.core.impl.SubjectBuilder;
 import org.opensaml.saml.saml2.core.impl.SubjectConfirmationBuilder;
 import org.opensaml.saml.saml2.core.impl.SubjectConfirmationDataBuilder;
 import org.wso2.carbon.identity.query.saml.exception.IdentitySAML2QueryException;
-import org.wso2.carbon.identity.query.saml.test.SPSignKeyDataHolder;
-import org.wso2.carbon.identity.query.saml.test.TestUtils;
-import org.wso2.carbon.identity.query.saml.util.OpenSAML3Util;
 import org.wso2.carbon.identity.query.saml.util.SAMLQueryRequestUtil;
+import org.wso2.carbon.identity.query.saml.util.OpenSAML3Util;
 
 import java.util.UUID;
 
 
-public class InvalidSAMLVersion {
+public class SAMLAttributeQueryRequestClient {
 
-    private final static Log log = LogFactory.getLog(InvalidSAMLVersion.class);
+    private final static Log log = LogFactory.getLog(SAMLAttributeQueryRequestClient.class);
 
     private static final String DIGEST_METHOD_ALGO = "http://www.w3.org/2000/09/xmldsig#rsa-sha1";
     private static final String SIGNING_ALGO = "http://www.w3.org/2000/09/xmldsig#sha1";
@@ -86,7 +84,7 @@ public class InvalidSAMLVersion {
 
         // AttributeQuery Request
         AttributeQuery attributeQuery = new AttributeQueryBuilder().buildObject();
-        attributeQuery.setVersion(SAMLVersion.VERSION_10);
+        attributeQuery.setVersion(SAMLVersion.VERSION_20);
         attributeQuery.setID(REQUEST_ID);
         attributeQuery.setIssueInstant(issueInstant);
         attributeQuery.setIssuer(issuer);
