+++ saml-query-profile-client/src/main/java/org/wso2/carbon/identity/saml/query/profile/test/invalid/InvalidSubject.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.query.saml.test.invalid;
+package org.wso2.carbon.identity.saml.query.profile.test.invalid;
 
 import org.apache.axiom.om.OMElement;
 import org.apache.axis2.client.ServiceClient;
@@ -25,40 +25,36 @@ import org.apache.commons.logging.LogFactory;
 import org.joda.time.DateTime;
 import org.opensaml.saml.common.SAMLVersion;
 import org.opensaml.saml.saml1.core.NameIdentifier;
-import org.opensaml.saml.saml2.core.AuthnContextComparisonTypeEnumeration;
-import org.opensaml.saml.saml2.core.AuthnQuery;
+import org.opensaml.saml.saml2.core.AttributeQuery;
 import org.opensaml.saml.saml2.core.Issuer;
 import org.opensaml.saml.saml2.core.NameID;
 import org.opensaml.saml.saml2.core.NameIDType;
-import org.opensaml.saml.saml2.core.RequestedAuthnContext;
 import org.opensaml.saml.saml2.core.Subject;
 import org.opensaml.saml.saml2.core.SubjectConfirmation;
 import org.opensaml.saml.saml2.core.SubjectConfirmationData;
-import org.opensaml.saml.saml2.core.impl.AuthnQueryBuilder;
+import org.opensaml.saml.saml2.core.impl.AttributeQueryBuilder;
 import org.opensaml.saml.saml2.core.impl.IssuerBuilder;
 import org.opensaml.saml.saml2.core.impl.NameIDBuilder;
-import org.opensaml.saml.saml2.core.impl.RequestedAuthnContextBuilder;
 import org.opensaml.saml.saml2.core.impl.SubjectBuilder;
 import org.opensaml.saml.saml2.core.impl.SubjectConfirmationBuilder;
 import org.opensaml.saml.saml2.core.impl.SubjectConfirmationDataBuilder;
 import org.wso2.carbon.identity.query.saml.exception.IdentitySAML2QueryException;
-import org.wso2.carbon.identity.query.saml.test.SPSignKeyDataHolder;
-import org.wso2.carbon.identity.query.saml.test.TestUtils;
+import org.wso2.carbon.identity.saml.query.profile.test.SPSignKeyDataHolder;
+import org.wso2.carbon.identity.saml.query.profile.test.TestUtils;
 import org.wso2.carbon.identity.query.saml.util.OpenSAML3Util;
 import org.wso2.carbon.identity.query.saml.util.SAMLQueryRequestUtil;
 
 import java.util.UUID;
 
 
-public class AuthnQuerySessionIndexClient {
+public class InvalidSubject {
 
-    private final static Log log = LogFactory.getLog(AuthnQuerySessionIndexClient.class);
+    private final static Log log = LogFactory.getLog(InvalidSubject.class);
 
     private static final String DIGEST_METHOD_ALGO = "http://www.w3.org/2000/09/xmldsig#rsa-sha1";
     private static final String SIGNING_ALGO = "http://www.w3.org/2000/09/xmldsig#sha1";
     private static final String ISSUER_ID = "travelocity.com";
-    private static final String NAME_ID = "admin";
-    private static final String SESSION_INDEX = "";
+    private static final String NAME_ID = "admin1";
 
     public static void main(String[] ags) throws Exception {
 
@@ -88,27 +84,23 @@ public class AuthnQuerySessionIndexClient {
         subject.getSubjectConfirmations().add(subjectConfirmation);
         subject.setNameID(nameID);
 
-        RequestedAuthnContext requestedAuthnContext = new RequestedAuthnContextBuilder().buildObject();
-        requestedAuthnContext.setComparison(AuthnContextComparisonTypeEnumeration.BETTER);
-
         // AttributeQuery Request
-        AuthnQuery authnQuery = new AuthnQueryBuilder().buildObject();
-        authnQuery.setVersion(SAMLVersion.VERSION_20);
-        authnQuery.setID(REQUEST_ID);
-        authnQuery.setIssueInstant(issueInstant);
-        authnQuery.setIssuer(issuer);
-        authnQuery.setSubject(subject);
-        authnQuery.setRequestedAuthnContext(requestedAuthnContext);
-        authnQuery.setSessionIndex(SESSION_INDEX);
+        AttributeQuery attributeQuery = new AttributeQueryBuilder().buildObject();
+        attributeQuery.setVersion(SAMLVersion.VERSION_20);
+        attributeQuery.setID(REQUEST_ID);
+        attributeQuery.setIssueInstant(issueInstant);
+        attributeQuery.setIssuer(issuer);
+        attributeQuery.setSubject(subject);
+        // End of AttributeQuery Request
 
         SAMLQueryRequestUtil.doBootstrap();
 
-        OpenSAML3Util.setSSOSignature(authnQuery, DIGEST_METHOD_ALGO,
+        OpenSAML3Util.setSSOSignature(attributeQuery, DIGEST_METHOD_ALGO,
                 SIGNING_ALGO, new SPSignKeyDataHolder());
 
         try {
-            body = SAMLQueryRequestUtil.marshall(authnQuery);
-            System.out.println("----Sample AuthnQuery Request Message----\n" + body);
+            body = SAMLQueryRequestUtil.marshall(attributeQuery);
+            System.out.println("----Sample AttributeQuery Request Message----\n" + body);
         } catch (Exception e) {
             log.error(e.getMessage());
             throw new IdentitySAML2QueryException("Error while marshalling the request.", e);
@@ -129,7 +121,6 @@ public class AuthnQuerySessionIndexClient {
          */
 
         ServiceClient serviceClient = TestUtils.createServiceClient();
-        // Set options to service client
         serviceClient.setOptions(TestUtils.setOptionsForServiceClient());
 
         // Set message to service.
@@ -137,7 +128,7 @@ public class AuthnQuerySessionIndexClient {
 
         // Printing return message.
         if (result != null) {
-            System.out.println("------Response Message From WSO2 Identity Server-----\n" + body);
+            System.out.println("------Response Message From WSO2 Identity Server-----\n" + result.toString());
         } else {
             log.error("Response message is null");
             throw new IdentitySAML2QueryException("Response message is null");
