+++ saml-query-profile-client/src/main/resources/wso2carbon.jks
